moved AboutUs component into components folder and updated import paths
