+++ src/pages/styles/AboutUs.css
@@ -18,20 +18,21 @@
     padding-top: 50px;
     padding-left: 150px;
   }
+
   .shadow-rectangle {
     margin-left: 100px;
   }
-  
+
 }
 
 @font-face {
   font-family: "JosefinSans";
-  src: url("../fonts/JosefinSans-VariableFont_wght.ttf") format("truetype");
+  src: url("./../../fonts/JosefinSans-VariableFont_wght.ttf") format("truetype");
 }
 
 @font-face {
   font-family: "Rastano";
-  src: url("../fonts/Rastano.ttf") format("truetype");
+  src: url("./../../fonts/Rastano.ttf") format("truetype");
 }
 
 .shadow-rectangle {
